Refactoring: rename FullscreenManager to DocumentFullscreen https://bugs.webkit.org/show_bug.cgi?id=259978 rdar://113633238
Reviewed by NOBODY (OOPS!).

FullscreenManager is an implementation of a IDL extension called Document+Fullscreen.
We already have a DocumentFullscreen.cpp/h file which calls out to FullscreenManager.
And so the name "FullscreenManager" is not strictly correct. It doesn't "manage" fullscreen;
it's merely a way to collect Document's fullscreen implementation into a separate class.

Rename FullscreenManager to DocumentFullscreen and move its implementation into the existing
DocumentFullscreen.cpp/h files. Rename Document::fullscreenManager() to just Document::fullscreen().

No functional changes here, just renames, so covered by existing testst.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/SelectorCheckerTestFunctions.h:
(WebCore::matchesWebkitFullScreenPseudoClass):
(WebCore::matchesFullScreenAnimatingFullScreenTransitionPseudoClass):
(WebCore::matchesFullScreenAncestorPseudoClass):
(WebCore::matchesFullScreenDocumentPseudoClass):
(WebCore::matchesFullScreenControlsHiddenPseudoClass):
* Source/WebCore/dom/Document.cpp:
(WebCore::Document::removedLastRef):
(WebCore::Document::commonTeardown):
* Source/WebCore/dom/Document.h:
(WebCore::Document::fullscreen):
(WebCore::Document::fullscreen const):
(WebCore::Document::fullscreenManager): Deleted.
(WebCore::Document::fullscreenManager const): Deleted.
* Source/WebCore/dom/DocumentFullscreen.cpp:
(WebCore::DocumentFullscreen::exitFullscreen):
(WebCore::DocumentFullscreen::webkitExitFullscreen):
(WebCore::DocumentFullscreen::fullscreenEnabled):
(WebCore::DocumentFullscreen::DocumentFullscreen):
(WebCore::DocumentFullscreen::fullscreenElement const):
(WebCore::DocumentFullscreen::requestFullscreenForElement):
(WebCore::DocumentFullscreen::cancelFullscreen):
(WebCore::documentsToUnfullscreen):
(WebCore::DocumentFullscreen::finishExitFullscreen):
(WebCore::DocumentFullscreen::isFullscreenEnabled const):
(WebCore::markRendererDirtyAfterTopLayerChange):
(WebCore::DocumentFullscreen::willEnterFullscreen):
(WebCore::DocumentFullscreen::didEnterFullscreen):
(WebCore::DocumentFullscreen::willExitFullscreen):
(WebCore::DocumentFullscreen::didExitFullscreen):
(WebCore::DocumentFullscreen::notifyAboutFullscreenChangeOrError):
(WebCore::DocumentFullscreen::dispatchEventForNode):
(WebCore::DocumentFullscreen::dispatchFullscreenChangeOrErrorEvent):
(WebCore::DocumentFullscreen::exitRemovedFullscreenElement):
(WebCore::DocumentFullscreen::isAnimatingFullscreen const):
(WebCore::DocumentFullscreen::setAnimatingFullscreen):
(WebCore::DocumentFullscreen::areFullscreenControlsHidden const):
(WebCore::DocumentFullscreen::setFullscreenControlsHidden):
(WebCore::DocumentFullscreen::clear):
(WebCore::DocumentFullscreen::emptyEventQueue):
(WebCore::DocumentFullscreen::addDocumentToFullscreenChangeEventQueue):
(WebCore::DocumentFullscreen::isSimpleFullscreenDocument const):
(WebCore::DocumentFullscreen::logChannel const):
* Source/WebCore/dom/DocumentFullscreen.h:
(WebCore::DocumentFullscreen::webkitFullscreenEnabled): Deleted.
(WebCore::DocumentFullscreen::webkitFullscreenElement): Deleted.
(WebCore::DocumentFullscreen::webkitIsFullScreen): Deleted.
(WebCore::DocumentFullscreen::webkitFullScreenKeyboardInputAllowed): Deleted.
(WebCore::DocumentFullscreen::webkitCurrentFullScreenElement): Deleted.
(WebCore::DocumentFullscreen::webkitCancelFullScreen): Deleted.
* Source/WebCore/dom/DocumentOrShadowRootFullscreen.cpp:
(WebCore::DocumentOrShadowRootFullscreen::fullscreenElement):
* Source/WebCore/dom/Element.cpp:
(WebCore::Element::removedFromAncestor):
(WebCore::Element::requestFullscreen):
* Source/WebCore/dom/EventPath.cpp:
* Source/WebCore/dom/FullscreenManager.cpp: Removed.
* Source/WebCore/dom/FullscreenManager.h: Removed.
* Source/WebCore/html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::elementIsHidden const):
(WebCore::HTMLMediaElement::isFullscreen const):
(WebCore::HTMLMediaElement::isStandardFullscreen const):
(WebCore::HTMLMediaElement::enterFullscreen):
(WebCore::HTMLMediaElement::exitFullscreen):
* Source/WebCore/html/MediaElementSession.cpp:
(WebCore::MediaElementSession::canShowControlsManager const):
(WebCore::MediaElementSession::updateMediaUsageIfChanged):
* Source/WebCore/html/shadow/MediaControlTextTrackContainerElement.cpp:
* Source/WebCore/inspector/agents/InspectorDOMAgent.cpp:
* Source/WebCore/page/EventHandler.cpp:
(WebCore::EventHandler::isKeyEventAllowedInFullScreen const):
(WebCore::EventHandler::internalKeyEvent):
* Source/WebCore/page/Page.cpp:
(WebCore::Page::setFullscreenControlsHidden):
* Source/WebCore/page/ScreenOrientation.cpp:
(WebCore::ScreenOrientation::lock):
* Source/WebCore/page/ios/ContentChangeObserver.cpp:
(WebCore::isHiddenBehindFullscreenElement):
* Source/WebCore/rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::RenderLayerBacking):
* Source/WebCore/rendering/RenderLayerCompositor.cpp:
(WebCore::isDescendantOfFullScreenLayer):
* Source/WebCore/rendering/RenderVideo.cpp:
(WebCore::RenderVideo::inElementOrVideoFullscreen const):
* Source/WebCore/style/StyleAdjuster.cpp:
(WebCore::Style::Adjuster::adjust const):
* Source/WebCore/style/StyleSharingResolver.cpp:
(WebCore::Style::SharingResolver::canShareStyleWithElement const):
* Source/WebCore/style/UserAgentStyle.cpp:
(WebCore::Style::UserAgentStyle::ensureDefaultStyleSheetsForElement):
* Source/WebCore/testing/Internals.cpp:
(WebCore::Internals::webkitWillEnterFullScreenForElement):
(WebCore::Internals::webkitDidEnterFullScreenForElement):
(WebCore::Internals::webkitWillExitFullScreenForElement):
(WebCore::Internals::webkitDidExitFullScreenForElement):
(WebCore::Internals::isAnimatingFullScreen const):
* Source/WebKit/WebProcess/FullScreen/WebFullScreenManager.cpp:
(WebKit::WebFullScreenManager::willEnterFullScreen):
(WebKit::WebFullScreenManager::didEnterFullScreen):
(WebKit::WebFullScreenManager::willExitFullScreen):
(WebKit::WebFullScreenManager::didExitFullScreen):
(WebKit::WebFullScreenManager::setAnimatingFullScreen):
(WebKit::WebFullScreenManager::requestRestoreFullScreen):
(WebKit::WebFullScreenManager::requestExitFullScreen):
(WebKit::WebFullScreenManager::handleEvent):
(WebKit::WebFullScreenManager::mainVideoElementTextRecognitionTimerFired):
* Source/WebKit/WebProcess/WebCoreSupport/WebChromeClient.cpp:
(WebKit::WebChromeClient::createWindow):
* Source/WebKit/WebProcess/WebCoreSupport/WebLocalFrameLoaderClient.cpp:
(WebKit::WebLocalFrameLoaderClient::dispatchDidStartProvisionalLoad):
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::elementDidFocus):
* Source/WebKitLegacy/mac/WebCoreSupport/WebChromeClient.mm:
(WebChromeClient::createWindow):
* Source/WebKitLegacy/mac/WebCoreSupport/WebKitFullScreenListener.mm:
(-[WebKitFullScreenListener webkitWillEnterFullScreen]):
(-[WebKitFullScreenListener webkitDidEnterFullScreen]):
(-[WebKitFullScreenListener webkitWillExitFullScreen]):
(-[WebKitFullScreenListener webkitDidExitFullScreen]):
* Source/WebKitLegacy/mac/WebView/WebFullScreenController.mm:
(-[WebFullScreenController requestExitFullScreen]):
(-[WebFullScreenController _manager]):
* Source/WebKitLegacy/mac/WebView/WebView.mm:
(-[WebView _didStartProvisionalLoadForFrame:]):
