+++ Source/WebCore/Sources.txt
@@ -1088,7 +1088,6 @@ dom/FocusEvent.cpp
 dom/FormDataEvent.cpp
 dom/FragmentDirectiveRangeFinder.cpp
 dom/FragmentDirectiveParser.cpp
-dom/FullscreenManager.cpp
 dom/GCReachableRef.cpp
 dom/HashChangeEvent.cpp
 dom/IdTargetObserver.cpp
