+++ Source/WebCore/html/MediaElementSession.cpp
@@ -31,10 +31,10 @@
 
 #include "Chrome.h"
 #include "ChromeClient.h"
+#include "DocumentFullscreen.h"
 #include "DocumentInlines.h"
 #include "DocumentLoader.h"
 #include "ElementInlines.h"
-#include "FullscreenManager.h"
 #include "HTMLAudioElement.h"
 #include "HTMLMediaElement.h"
 #include "HTMLNames.h"
@@ -617,7 +617,7 @@ bool MediaElementSession::canShowControlsManager(PlaybackControlsPurpose purpose
 
 #if ENABLE(FULLSCREEN_API)
     // Elements which are not descendants of the current fullscreen element cannot be main content.
-    auto* fullscreenElement = m_element.document().fullscreenManager().currentFullscreenElement();
+    auto* fullscreenElement = m_element.document().fullscreen().currentFullscreenElement();
     if (fullscreenElement && !m_element.isDescendantOf(*fullscreenElement)) {
         INFO_LOG(LOGIDENTIFIER, "returning FALSE: outside of full screen");
         return false;
@@ -1307,7 +1307,7 @@ void MediaElementSession::updateMediaUsageIfChanged()
 
     bool isOutsideOfFullscreen = false;
 #if ENABLE(FULLSCREEN_API)
-    if (auto* fullscreenElement = document.fullscreenManager().currentFullscreenElement())
+    if (auto* fullscreenElement = document.fullscreen().currentFullscreenElement())
         isOutsideOfFullscreen = !m_element.isDescendantOf(*fullscreenElement);
 #endif
     bool isAudio = client().presentationType() == MediaType::Audio;
